Restructured BedAnnotations, general parseRecords
I made use of declaration merging and the Omit<type, keys> type to
get unique classes for BED3, 6, 9, and 12. It should work pretty well,
but it'll need some testing.

I implemented a simple parseRecordsFromString() function that can be
used to parse multiple records from a string. It's typed with an
Annotation class and takes a reference to a parsing function that
returns that type.