Restructured Chart rendering callbacks
The Chart rendering callbacks are now typed with (this: any, params: P),
but are set to readonly. The callback parameters in ChartConfig are
typed with (this: Chart<P>, params: P). This allows a proper this
context, should avoid the contravariance issues for classes that inherit from
Chart, and still enforces type safety.