Fixed some bugs in AnnotationGroup
It turns out that I needed to check for group.length == 0 instead of
group == [] for the add condition on AnnotationGroups.

It also seems like I needed to define a getter for y since I overwrote
the base Annotation setter for y. This seems odd to me, but so be it.