Fixed selector based re-rendering bug
When providing an explicit selector in a glyph call, repeated calls
using the same selector is supposed to result in replacing the glyphs
from the original call with the ones from the new call. This was not
working after the glyph binding rework that organized everything more
neatly in g elements.

This commit reorganizes the bind function to appropriately assign css
classes to the g tags in a way that made it possible to restore the
re-rendering functionality.