Annotation is now an interface
I've been planning to do this for a while--it makes SODA more
generalizable.

This required lots of small changes to most of the codebase. Most
features survived the changes, but there are definitely some things
broken that I'll have to sort out. Plot glyphs are in a bad state,
the layout code is going to need to change substantially, and probably
more that I haven't quite identified yet.