Changed TS module to ESNext
I think this is what we want for the module setting. I haven't changed
the target, so it shouldn't mess with browser compatibility (if I
understand this correctly).

I also added module exports and stuff to the package.json. I'm hoping
that these two things will fix the skypack bundling.