Fixed dynamic text bug
The dynamic text glyph was broken in two ways:

(i) when I rewrote it, I used Chart.range for the threshold/ratio
calculations, when I should have used Chart.domain. I fixed that.

(ii) when I added in more text configuration a long time ago, I failed
to consider that I needed to include those configurations in the
function that precomputes the size of each string. Fixed that.