Reworked glyph queries
Glyph queries were super inefficient when each query field was not
specified. This isn't a crazy speed increase, but it should help a bit.

It changes the GlyphQueryConfig to use a list of Annotations rather than
an ID. I had to change the way click and hover behaviors work internally
to support the changes to glyph queries.